+++ zarrs_storage/src/storage_async.rs
@@ -222,6 +222,12 @@ pub trait AsyncWritableStorageTraits: MaybeSend + MaybeSync {
     /// # Errors
     /// Returns a [`StorageError`] if there is an underlying storage error.
     async fn erase_prefix(&self, prefix: &StorePrefix) -> Result<(), StorageError>;
+
+    /// Returns whether this store supports partial writes.
+    ///
+    /// If this returns `true`, the store can efficiently handle `set_partial` and `set_partial_many` operations.
+    /// If this returns `false`, partial sets will fall back to a full read and write operation.
+    fn supports_set_partial(&self) -> bool;
 }
 
 /// A supertrait of [`AsyncReadableStorageTraits`] and [`AsyncWritableStorageTraits`].